Rename version file with git tag version
