+++ packages/shorebird_cli/lib/src/commands/upgrade_command.dart
@@ -5,6 +5,7 @@ import 'package:path/path.dart' as p;
 import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/shorebird_environment.dart';
+import 'package:shorebird_cli/src/git_tag_version.dart';
 import 'package:shorebird_cli/src/shorebird_version_mixin.dart';
 
 /// {@template upgrade_command}
