+++ packages/shorebird_cli/lib/src/git_tag_version.dart
