+++ packages/shorebird_cli/test/src/commands/upgrade_command_test.dart
@@ -3,7 +3,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/commands/commands.dart';
 import 'package:shorebird_cli/src/logger.dart';
-import 'package:shorebird_cli/src/commands/version.dart';
+import 'package:shorebird_cli/src/git_tag_version.dart';
 import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:test/test.dart';
